+++ rust/processor/src/db/common/models/ans_models/mod.rs
@@ -1,4 +1,2 @@
 pub mod raw_ans_lookup_v2;
 pub mod raw_ans_primary_name_v2;
-pub mod raw_current_ans_lookup_v2;
-pub mod raw_current_ans_primary_name_v2;
