+++ rust/processor/src/db/postgres/models/ans_models/mod.rs
@@ -3,6 +3,7 @@
 
 pub mod ans_lookup;
 pub mod ans_lookup_v2;
+pub mod ans_primary_name_v2;
 pub mod ans_utils;
 
 // parquet models
