+++ rust/processor/src/processors/ans_processor.rs
@@ -3,12 +3,23 @@
 
 use super::{DefaultProcessingResult, ProcessorName, ProcessorTrait};
 use crate::{
-    db::postgres::models::ans_models::{
-        ans_lookup::{AnsLookup, AnsPrimaryName, CurrentAnsLookup, CurrentAnsPrimaryName},
-        ans_lookup_v2::{
-            AnsLookupV2, AnsPrimaryNameV2, CurrentAnsLookupV2, CurrentAnsPrimaryNameV2,
+    db::{
+        common::models::ans_models::{
+            raw_ans_lookup_v2::{
+                AnsLookupV2Convertible, CurrentAnsLookupV2Convertible, RawAnsLookupV2,
+                RawCurrentAnsLookupV2,
+            },
+            raw_ans_primary_name_v2::{
+                AnsPrimaryNameV2Convertible, CurrentAnsPrimaryNameV2Convertible,
+                RawAnsPrimaryNameV2, RawCurrentAnsPrimaryNameV2,
+            },
         },
-        ans_utils::{RenewNameEvent, SubdomainExtV2},
+        postgres::models::ans_models::{
+            ans_lookup::{AnsLookup, AnsPrimaryName, CurrentAnsLookup, CurrentAnsPrimaryName},
+            ans_lookup_v2::{AnsLookupV2, CurrentAnsLookupV2},
+            ans_primary_name_v2::{AnsPrimaryNameV2, CurrentAnsPrimaryNameV2},
+            ans_utils::{RenewNameEvent, SubdomainExtV2},
+        },
     },
     gap_detectors::ProcessingResult,
     schema,
@@ -391,6 +402,27 @@ impl ProcessorTrait for AnsProcessor {
             self.config.ans_v2_contract_address.clone(),
         );
 
+        let postgres_current_ans_lookup_v2: Vec<CurrentAnsLookupV2> = all_current_ans_lookups_v2
+            .iter()
+            .map(CurrentAnsLookupV2::from_raw)
+            .collect();
+
+        let postgres_ans_lookup_v2: Vec<AnsLookupV2> = all_ans_lookups_v2
+            .iter()
+            .map(AnsLookupV2::from_raw)
+            .collect();
+
+        let postgres_current_ans_primary_name_v2: Vec<CurrentAnsPrimaryNameV2> =
+            all_current_ans_primary_names_v2
+                .iter()
+                .map(CurrentAnsPrimaryNameV2::from_raw)
+                .collect();
+
+        let postgres_ans_primary_name_v2: Vec<AnsPrimaryNameV2> = all_ans_primary_names_v2
+            .iter()
+            .map(AnsPrimaryNameV2::from_raw)
+            .collect();
+
         let processing_duration_in_secs = processing_start.elapsed().as_secs_f64();
         let db_insertion_start = std::time::Instant::now();
 
@@ -432,10 +464,10 @@ impl ProcessorTrait for AnsProcessor {
             &all_ans_lookups,
             &all_current_ans_primary_names,
             &all_ans_primary_names,
-            &all_current_ans_lookups_v2,
-            &all_ans_lookups_v2,
-            &all_current_ans_primary_names_v2,
-            &all_ans_primary_names_v2,
+            &postgres_current_ans_lookup_v2,
+            &postgres_ans_lookup_v2,
+            &postgres_current_ans_primary_name_v2,
+            &postgres_ans_primary_name_v2,
             &self.per_table_chunk_sizes,
         )
         .await;
@@ -480,10 +512,10 @@ pub fn parse_ans(
     Vec<AnsLookup>,
     Vec<CurrentAnsPrimaryName>,
     Vec<AnsPrimaryName>,
-    Vec<CurrentAnsLookupV2>,
-    Vec<AnsLookupV2>,
-    Vec<CurrentAnsPrimaryNameV2>,
-    Vec<AnsPrimaryNameV2>,
+    Vec<RawCurrentAnsLookupV2>,
+    Vec<RawAnsLookupV2>,
+    Vec<RawCurrentAnsPrimaryNameV2>,
+    Vec<RawAnsPrimaryNameV2>,
 ) {
     let mut all_current_ans_lookups = AHashMap::new();
     let mut all_ans_lookups = vec![];
@@ -532,7 +564,7 @@ pub fn parse_ans(
                     v2_renew_name_events.push(renew_name_event);
                 }
                 if let Some((current_ans_lookup_v2, ans_lookup_v2)) =
-                    CurrentAnsPrimaryNameV2::parse_v2_primary_name_record_from_event(
+                    RawCurrentAnsPrimaryNameV2::parse_v2_primary_name_record_from_event(
                         event,
                         txn_version,
                         event_index as i64,
@@ -595,7 +627,10 @@ pub fn parse_ans(
 
                             // Include all v1 lookups in v2 data
                             let (current_ans_lookup_v2, ans_lookup_v2) =
-                                CurrentAnsLookupV2::get_v2_from_v1(current_ans_lookup, ans_lookup);
+                                RawCurrentAnsLookupV2::get_v2_from_v1(
+                                    current_ans_lookup,
+                                    ans_lookup,
+                                );
                             all_current_ans_lookups_v2
                                 .insert(current_ans_lookup_v2.pk(), current_ans_lookup_v2);
                             all_ans_lookups_v2.push(ans_lookup_v2);
@@ -622,7 +657,7 @@ pub fn parse_ans(
 
                             // Include all v1 primary names in v2 data
                             let (current_primary_name_v2, primary_name_v2) =
-                                CurrentAnsPrimaryNameV2::get_v2_from_v1(current_primary_name.clone(), primary_name.clone());
+                                RawCurrentAnsPrimaryNameV2::get_v2_from_v1(current_primary_name.clone(), primary_name.clone());
                             all_current_ans_primary_names_v2
                                 .insert(current_primary_name_v2.pk(), current_primary_name_v2);
                             all_ans_primary_names_v2.push(primary_name_v2);
@@ -650,7 +685,10 @@ pub fn parse_ans(
 
                             // Include all v1 lookups in v2 data
                             let (current_ans_lookup_v2, ans_lookup_v2) =
-                                CurrentAnsLookupV2::get_v2_from_v1(current_ans_lookup, ans_lookup);
+                                RawCurrentAnsLookupV2::get_v2_from_v1(
+                                    current_ans_lookup,
+                                    ans_lookup,
+                                );
                             all_current_ans_lookups_v2
                                 .insert(current_ans_lookup_v2.pk(), current_ans_lookup_v2);
                             all_ans_lookups_v2.push(ans_lookup_v2);
@@ -676,7 +714,7 @@ pub fn parse_ans(
 
                             // Include all v1 primary names in v2 data
                             let (current_primary_name_v2, primary_name_v2) =
-                                CurrentAnsPrimaryNameV2::get_v2_from_v1(current_primary_name, primary_name);
+                                RawCurrentAnsPrimaryNameV2::get_v2_from_v1(current_primary_name, primary_name);
                             all_current_ans_primary_names_v2
                                 .insert(current_primary_name_v2.pk(), current_primary_name_v2);
                             all_ans_primary_names_v2.push(primary_name_v2);
@@ -684,7 +722,7 @@ pub fn parse_ans(
                     },
                     WriteSetChange::WriteResource(write_resource) => {
                         if let Some((current_ans_lookup_v2, ans_lookup_v2)) =
-                            CurrentAnsLookupV2::parse_name_record_from_write_resource_v2(
+                            RawCurrentAnsLookupV2::parse_name_record_from_write_resource_v2(
                                 write_resource,
                                 &ans_v2_contract_address,
                                 txn_version,
@@ -722,10 +760,10 @@ pub fn parse_ans(
         .collect::<Vec<CurrentAnsPrimaryName>>();
     let mut all_current_ans_lookups_v2 = all_current_ans_lookups_v2
         .into_values()
-        .collect::<Vec<CurrentAnsLookupV2>>();
+        .collect::<Vec<RawCurrentAnsLookupV2>>();
     let mut all_current_ans_primary_names_v2 = all_current_ans_primary_names_v2
         .into_values()
-        .collect::<Vec<CurrentAnsPrimaryNameV2>>();
+        .collect::<Vec<RawCurrentAnsPrimaryNameV2>>();
 
     all_current_ans_lookups.sort();
     all_current_ans_primary_names.sort();
